requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
python-dotenv==1.2.1
pydantic==2.12.5
slowapi==0.1.9